+++ app/Http/Controllers/Api/V1/UsersController.php
@@ -4,12 +4,9 @@ namespace App\Http\Controllers\Api\V1;
 
 use App\Http\Controllers\Controller;
 use App\Http\Traits\SearchTrait;
-use App\Models\Company;
 use App\Models\Group;
-use App\Models\User;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\DB;
-use Illuminate\Validation\Rule;
 
 class UsersController extends Controller
 {
@@ -56,102 +53,6 @@ class UsersController extends Controller
         return $this->model->get();
     }
 
-    public function store()
-    {
-        $data = $this->filterData(request()->all());
-        $validator = Validator::make($data, $this->getRules(), $this->getMessages());
-
-        if ($validator->fails()) {
-            return response()->json($validator->errors(), 422);
-        }
-
-        return User::create($data);
-    }
-
-    public function show(User $user)
-    {
-        return $user;
-    }
-
-    public function update(User $user)
-    {
-        $data = $this->filterData(request()->all());
-        $validator = Validator::make($data, $this->getRules($user), $this->getMessages());
-
-        if ($validator->fails()) {
-            return response()->json($validator->errors(), 422);
-        }
-
-        if (!$data) {
-            return response()->json(['message' => 'É necessário enviar parametros!'], 422);
-        }
-
-        $user->update($data);
-        return $user;
-    }
-
-    public function destroy(User $user)
-    {
-        $user->delete();
-        return response()->json(['result' => 'ok']);
-    }
-
-    private function filterData($inputs)
-    {
-        $data = [];
-        $list = array_filter((array) $inputs, 'strlen');
-
-        foreach ($list as $key => $value) {
-            $data[$key] = trim(strip_tags($value));
-        }
-
-        if (isset($data['cpf_cnpj'])) {
-            $data['cpf_cnpj'] = preg_replace('/[^\d\,]/', '', $data['cpf_cnpj']);
-        }
-
-        if (isset($data['group'])) {
-            $data['group'] = preg_replace('/[^\d\,]/', '', $data['group']);
-        }
-
-        return $data;
-    }
-
-    public function getRules(User $user = null)
-    {
-        if ($user) {
-            $rules = [
-                'group_id' => 'exists:groups,id',
-                'name' => 'max:100',
-                'cpf_cnpj' => ['cpf_cnpj', Rule::unique('users')->ignore($user)],
-                'email' => ['email', Rule::unique('users')->ignore($user)],
-                'username' => 'max:100',
-            ];
-        } else {
-            $rules = [
-                'group_id' => 'required|exists:groups,id',
-                'name' => 'required|max:100',
-                'cpf_cnpj' => 'required|cpf_cnpj|unique:users',
-                'email' => 'required|email|unique:users',
-                'username' => 'required|max:100',
-                'password' => 'required',
-            ];
-        }
-
-        return $rules;
-    }
-
-    public function getMessages()
-    {
-        return [
-            'required' => 'O campo é obrigatório!',
-            'exists' => 'Não existe esse grupo.',
-            'cpf_cnpj' => 'CPF/CNPJ é inválido!',
-            'max' => 'Máximo de :max caracteres!',
-            'unique' => ':input já em uso!',
-            'email' => 'E-mail inválido!'
-        ];
-    }
-
     public function storeBatches()
     {
         $inputs = request()->input('users');
@@ -178,71 +79,6 @@ class UsersController extends Controller
             $data = $this->filterData($input);
             $group = Group::where('cnpj', $data['group'])->first();
             $group->users()->create($data);
-            /*
-            User::create(array_merge($data, ['group_id' => $group->id]));
-            */
-        }
-
-        return response()->json(['result' => 'ok']);
-    }
-
-    public function updateBatches()
-    {
-        $inputs = request()->input('users');
-
-        $errors = [];
-        foreach ($inputs as $input) {
-            $data = $this->filterData($input);
-
-            $validator = Validator::make($data, $this->getRulesUpdateAndDestroyBatches(), $this->getMessages());
-
-            if ($validator->fails()) {
-                $errors[] = [
-                    'fields' => $input,
-                    'errors' => $validator->errors()
-                ];
-            }
-        }
-
-        if (count($errors)) {
-            return response()->json($errors, 422);
-        }
-
-        foreach ($inputs as $input) {
-            $data = $this->filterData($input);
-            $user = User::where('cpf_cnpj', $data['cpf_cnpj'])->first();
-            $user->update($data);
-        }
-
-        return response()->json(['result' => 'ok']);
-    }
-
-    public function destroyBatches()
-    {
-        $inputs = request()->input('users');
-
-        $errors = [];
-        foreach ($inputs as $input) {
-            $data = $this->filterData($input);
-
-            $validator = Validator::make($data, $this->getRulesUpdateAndDestroyBatches(true), $this->getMessages());
-
-            if ($validator->fails()) {
-                $errors[] = [
-                    'fields' => $input,
-                    'errors' => $validator->errors()
-                ];
-            }
-        }
-
-        if (count($errors)) {
-            return response()->json($errors, 422);
-        }
-
-        foreach ($inputs as $input) {
-            $data = $this->filterData($input);
-            $user = User::where('cpf_cnpj', $data['cpf_cnpj'])->first();
-            $user->delete();
         }
 
         return response()->json(['result' => 'ok']);
@@ -250,12 +86,55 @@ class UsersController extends Controller
 
     public function destroyAll()
     {
-        $cnpj = request()->input('group');
-        $group = Group::where('cnpj', $cnpj)->first();
+        $data = request()->all();
+
+        if (!isset($data['group'])) {
+            return response()->json(['message' => 'É necessario enviar o grupo!'], 422);
+        }
+
+        $group = Group::where('cnpj', preg_replace('/[^\d\,]/', '', $data['group']))->first();
+
+        if (!$group) {
+            return response()->json(['message' => 'Grupo não encontrado!'], 422);
+        }
+
         foreach ($group->users as $user) {
             $user->delete();
         }
+
         return response()->json(['result' => 'ok']);
+    }
+
+    private function filterData($inputs)
+    {
+        $data = [];
+        $list = array_filter((array) $inputs, 'strlen');
+
+        foreach ($list as $key => $value) {
+            $data[$key] = trim(strip_tags($value));
+        }
+
+        if (isset($data['cpf_cnpj'])) {
+            $data['cpf_cnpj'] = preg_replace('/[^\d\,]/', '', $data['cpf_cnpj']);
+        }
+
+        if (isset($data['group'])) {
+            $data['group'] = preg_replace('/[^\d\,]/', '', $data['group']);
+        }
+
+        return $data;
+    }
+
+    public function getMessages()
+    {
+        return [
+            'required' => 'O campo é obrigatório!',
+            'exists' => 'Não existe esse grupo.',
+            'cpf_cnpj' => 'CPF/CNPJ é inválido!',
+            'max' => 'Máximo de :max caracteres!',
+            'unique' => ':input já em uso!',
+            'email' => 'E-mail inválido!'
+        ];
     }
 
     public function getRulesStoreBatches()
@@ -268,20 +147,5 @@ class UsersController extends Controller
             'username' => 'required|max:100',
             'password' => 'required',
         ];
-    }
-
-    public function getRulesUpdateAndDestroyBatches($destroy = false)
-    {
-        $rules = $destroy ? [
-            'group' => 'required|exists:groups,cnpj',
-        ] : [
-            'group' => 'required|exists:groups,cnpj',
-            'name' => 'required|max:100',
-            'cpf_cnpj' => 'required|cpf_cnpj|exists:users,cpf_cnpj',
-            'email' => 'required|email|exists:users,email',
-            'username' => 'required|max:100',
-            'password' => 'required',
-        ];
-        return $rules;
     }
 }
